e2e/pyproject.toml
[tool.poetry]
name = "e2e"
version = "0.1.0"
description = ""
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.2.2"
pytest-playwright = "^0.5.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
